[refactor] Rename invoice to heroes
