+++ app/dashboard/heroes/[id]/edit/not-found.js
@@ -10,7 +10,7 @@ export default function NotFound() {
     <main className="flex h-full flex-col items-center justify-center gap-2">
       <FaceFrownIcon className="w-10 text-gray-400" />
       <h2 className="text-xl font-semibold">404 Not Found</h2>
-      <p>Could not find the requested invoice.</p>
+      <p>Could not find the requested Hero.</p>
       <Link
         href="/dashboard/heroes"
         className="mt-4 rounded-md bg-blue-500 px-4 py-2 text-sm text-white transition-colors hover:bg-blue-400"
