+++ app/dashboard/heroes/error.js
@@ -17,7 +17,7 @@ export default function Error({ error, reset }) {
       <button
         className="mt-4 rounded-md bg-blue-500 px-4 py-2 text-sm text-white transition-colors hover:bg-blue-400"
         onClick={
-          // Attempt to recover by trying to re-render the invoices route
+          // Attempt to recover by trying to re-render the heroes route
           () => reset()
         }
       >
