+++ app/lib/data.ts
@@ -33,7 +33,7 @@ export async function fetchLatestHeroUploads() {
     return data.rows;
   } catch (error) {
     console.error("Database Error:", error);
-    throw new Error("Failed to fetch the latest invoices.");
+    throw new Error("Failed to fetch the latest heroes.");
   }
 }
